fix: test and fix pages

+++ app/Http/Controllers/Back/PageController.php
@@ -11,13 +11,10 @@ use Illuminate\Http\Request;
 
 class PageController extends Controller
 {
-    public function __construct()
-    {
-        $this->authorizeResource(Page::class, 'page');
-    }
-
     public function index()
     {
+        $this->authorize('pages.index');
+
         $pages = Page::latest()->paginate(10);
 
         return view('back.pages.index', compact('pages'));
@@ -25,11 +22,15 @@ class PageController extends Controller
 
     public function create()
     {
+        $this->authorize('pages.create');
+
         return view('back.pages.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('pages.create');
+
         $this->validate($request, [
             'title' => 'required|string|max:191',
             'content' => 'required',
@@ -37,10 +38,10 @@ class PageController extends Controller
         ]);
 
         Page::create([
-            'title'      => $request->title,
-            'content'    => $request->content,
-            'slug'       => $request->slug ?: $request->title,
-            'published'  => $request->published ? true : false,
+            'title' => $request->title,
+            'content' => $request->input('content'),
+            'slug' => SlugService::createSlug(Page::class, 'slug', $request->slug ?: $request->title),
+            'published' => (bool)$request->published,
         ]);
 
         toastr()->success('صفحه با موفقیت ایجاد شد.');
@@ -50,31 +51,35 @@ class PageController extends Controller
 
     public function edit(Page $page)
     {
+        $this->authorize('pages.update');
+
         return view('back.pages.edit', compact('page'));
     }
 
     public function update(Request $request, Page $page)
     {
+        $this->authorize('pages.update');
+
         $this->validate($request, [
             'title' => 'required|string|max:191',
             'content' => 'required',
         ]);
 
-        $slug = $page->slug;
+        $slug = SlugService::createSlug(Page::class, 'slug', $request->slug ?: $request->title);
+
+        Menu::where('url', $page->link($page->slug))->update([
+            'url' => $page->link($slug),
+        ]);
+
+        Link::where('url', $page->link($page->slug))->update([
+            'url' => $page->link($slug),
+        ]);
 
         $page->update([
-            'title'     => $request->title,
-            'content'   => $request->content,
-            'slug'      => $request->slug ?: $request->title,
-            'published' => $request->published ? true : false,
-        ]);
-
-        Menu::where('link', '/pages/' . $slug)->update([
-            'link' => '/pages/' . $page->slug,
-        ]);
-
-        Link::where('link', '/pages/' . $slug)->update([
-            'link' => '/pages/' . $page->slug,
+            'title' => $request->title,
+            'content' => $request->input('content'),
+            'slug' => $slug,
+            'published' => (bool)$request->published,
         ]);
 
         toastr()->success('صفحه با موفقیت ویرایش شد.');
@@ -84,6 +89,8 @@ class PageController extends Controller
 
     public function destroy(Page $page)
     {
+        $this->authorize('pages.delete');
+
         $page->delete();
 
         return response("success", 200);
